wait_for_event() helper
